feat(davinci): one projection mapping model for Canon and Maestro (P4-5a slice 1)
`vize_canon::virtual_ts::mapping` now owns `ProjectionMapping`, the single
span-link model of charter #14: the generators' `VizeMapping` rows plus a
per-row `ProjectionMeta` column (feature flags and construct kind), semantic
links and an authored base. Canon's `source_map` module and Maestro's
virtual-code source map are deleted; every Maestro `VirtualDocument`,
`TypeIntelligence`, `SfcSourceMap` (now a thin view) and Corsa's
`TypeCheckResult` read the one model.

Caret lookups keep Maestro's byte-to-byte arithmetic and diagnostic lookups
keep the narrowest-row/sub-span arithmetic both checker paths share. The TS-40
Maestro record renders its frozen spelling from the model, and all 13 TS-40
snapshots (default and legacy lanes) are byte-identical.

+++ crates/vize_maestro/src/ide/corsa_support/art_variant_fallback_tests.rs
@@ -82,7 +82,7 @@ fn art_variant_hover_answers_from_the_authored_template_without_a_generated_offs
         .and_then(|documents| documents.art_template(info.variant_index))
         .expect("typed art template");
     assert!(
-        template.source_map.to_generated(offset as u32).is_none(),
+        template.source_map.to_generated(offset).is_none(),
         "fixture must exercise the unmapped variant offset",
     );
